backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.42.0
langchain==0.3.14
langchain-anthropic==0.3.4
langchain-community==0.3.14
langchain-text-splitters==0.3.4
faiss-cpu==1.9.0.post1
python-multipart==0.0.20
sse-starlette==2.2.1
